Adds Danger to automate common code review chores
We would like to have good GIT hygiene in this repository including well
formatted commit messages. Danger is a great system to automate common
code review chores such as this.

In order to get Danger working in the pipeline this adds a
.tool-versions for the ASDF version manager to lock versions of
languages (Ruby in this case for Danger).

This adds Danger and supporting files. The initial goal is to lint
commit messages.

See: https://danger.systems/ruby/